Add Faction Names

+++ core/src/main/java/heroes/journey/components/StatsComponent.java
@@ -1,18 +1,17 @@
 package heroes.journey.components;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import com.badlogic.ashley.core.ComponentMapper;
 import com.badlogic.ashley.core.Entity;
-
 import heroes.journey.components.interfaces.ClonableComponent;
 import heroes.journey.entities.stats.DamageType;
 import heroes.journey.entities.stats.DamageTypeManager;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class StatsComponent implements ClonableComponent<StatsComponent> {
 
-    public static final float MAX_HEALTH = 100, MAX_MANA = 100;
+    public static final int MAX_HEALTH = 100, MAX_MANA = 100;
 
     private int body, mind;
     private int fame;
@@ -20,21 +19,21 @@ public class StatsComponent implements ClonableComponent<StatsComponent> {
     private int health;
     private int mana;
 
-    private final Map<DamageType,Integer> resistancesFlat;
-    private final Map<DamageType,Float> resistancesPercentage;
+    private final Map<DamageType, Integer> resistancesFlat;
+    private final Map<DamageType, Float> resistancesPercentage;
 
     public StatsComponent() {
         body = 1;
         mind = 1;
         fame = 0;
-        resistancesFlat = new HashMap<DamageType,Integer>();
-        resistancesPercentage = new HashMap<DamageType,Float>();
+        resistancesFlat = new HashMap<DamageType, Integer>();
+        resistancesPercentage = new HashMap<DamageType, Float>();
         for (DamageType type : DamageTypeManager.get().values()) {
             resistancesFlat.put(type, 0);
             resistancesPercentage.put(type, 0f);
         }
-        health = (int)MAX_HEALTH;
-        mana = (int)MAX_MANA;
+        health = (int) MAX_HEALTH;
+        mana = (int) MAX_MANA;
     }
 
     public int getBody() {
@@ -67,7 +66,7 @@ public class StatsComponent implements ClonableComponent<StatsComponent> {
 
     // Returns if they are Alive
     public boolean adjustHealth(int health) {
-        this.health = (int)Math.min(MAX_HEALTH, Math.max(0, this.health + health));
+        this.health = (int) Math.min(MAX_HEALTH, Math.max(0, this.health + health));
         return (this.health + health > 0);
     }
 
@@ -80,15 +79,15 @@ public class StatsComponent implements ClonableComponent<StatsComponent> {
         if (this.mana + mana < 0) {
             return false;
         }
-        this.mana = (int)Math.min(MAX_MANA, Math.max(0, this.mana + mana));
+        this.mana = (int) Math.min(MAX_MANA, Math.max(0, this.mana + mana));
         return true;
     }
 
-    public Map<DamageType,Integer> getResistancesFlat() {
+    public Map<DamageType, Integer> getResistancesFlat() {
         return resistancesFlat;
     }
 
-    public Map<DamageType,Float> getResistancesPercentage() {
+    public Map<DamageType, Float> getResistancesPercentage() {
         return resistancesPercentage;
     }
